+++ test-parts/part-h-api_requests/api-main.js
@@ -1,8 +1,8 @@
 const cAdminFile = require("./requests/req-admin_main");
 const cHealthFile = require("./requests/req-admin_health");
 const cAlarmFile = require("./requests/req-alarm");
-/*
 const cStorageFile = require("./requests/req-storage");
+/*
 const cDevicesFileCrud = require("./requests/req-devices_crud");
 const cDevicesFileCrudInvalid = require("./requests/req-devices_crud_invalid");
 const cDevicesFileModifyInvalid = require("./requests/req-devices_modify_invalid");
@@ -15,8 +15,8 @@ function coordinateApiRequests()
 		cAdminFile.callTestAdminApis();
 		cHealthFile.callTestHealthApi();
 		cAlarmFile.callTestAlarmApis();
-		/*
 		cStorageFile.callTestStorageAPIs();
+		/*
 		cDevicesFileCrud.callTestDeviceCrudApis();
 		cDevicesFileCrudInvalid.callTestDeviceCrudInvalidApis();
 		cDevicesFileModifyInvalid.callTestDeviceModifyInvalidApis();